Move config to a `step` package
